Fixed tests, moved table creation code from evaka ops file
